moved general popup functionality to a base class
